fix(session-store): bound native ownership and preserve cold witnesses
Preserve native snapshot export and full-key shared row maps, reduce
retained metadata and temporary catalog allocations, and keep exact
lineage, reservations, retention and publication checks. Route native
roster observations through their actual owner and schedule large
evaluation/publication work under independent bounded gates.

Keep backend-independent quorum controls, original follower cuts and
deadlines, cold reconstruction witnesses, and owned process-loss checks.
Move the existing snapshot extraction buffer off nested startup frames
and integrate the reviewed Clippy cleanup with allocation diagnostics.

Focused native/SQL, quorum, cold snapshot, maximum-body and process-loss
checks passed; the combined package is Clippy clean. Full-scale RSS and
remaining repository/hosted qualification are still open on the draft PR.

+++ crates/opc-session-store/src/sqlite/consensus/tests/native/basis/cold_reads.rs
@@ -206,7 +206,7 @@ fn native_selected_log_read_returns_eight_full_256_request_rows_with_shared_memo
             let requests = (0..count)
                 .map(|slot| sdk741_component_request(Sdk741Payload::Create, 10 + batch, slot, None))
                 .collect::<Vec<_>>();
-            fenced_transition_v2_batch_entry(2 + batch as u64, requests, timestamp(2 + batch as u8))
+            fenced_transition_v2_batch_entry(2 + batch, requests, timestamp(2 + batch as u8))
         })
         .collect::<Vec<_>>();
     fixture.parity(&entries);
